requirements.txt
anyio==4.0.0
async==0.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pyTelegramBotAPI==4.13.0
python-telegram-bot==20.5
requests==2.31.0
sniffio==1.3.0
telebot==0.0.5
urllib3==2.0.4
